HBASE-8802 totalCompactingKVs overflow (Chao Shi)
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.95@1497051 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/compactions/Compactor.java
@@ -108,7 +108,7 @@ public abstract class Compactor {
   /** The sole reason this class exists is that java has no ref/out/pointer parameters. */
   protected static class FileDetails {
     /** Maximum key count after compaction (for blooms) */
-    public int maxKeyCount = 0;
+    public long maxKeyCount = 0;
     /** Earliest put timestamp if major compaction */
     public long earliestPutTs = HConstants.LATEST_TIMESTAMP;
     /** The last key in the files we're compacting. */
